Update loading component
