+++ src/components/common/index.ts
@@ -3,3 +3,4 @@ export * from './EmptyListCTA';
 export * from './Pagination';
 export * from './SecretInput';
 export * from './Page';
+export * from './Loading';
